Changed the communication cycle
Previously, the family information was being provided via the response
from the api endpoints. this has been replaced to better adhere to the
mermaid diagram. Now, the signup and login simply inform of their userID
and the success/error of the action. It also starts a chain of events
which results in the data being requested via the socket instead of
being provided as a repsonse to the api request