New Member in FamilyView
Refactored Tree.tsx to resolve the cyclical nature of rendering this
tree

Changed the code so that we can indicate that a node has already been
visited and rendered. This avoids entering an infinite recursion loop
when rendering the tree

Changed how the Authentication and NewMember are writing the family data
to AsyncStorage. This is because it needs to be flatted first, before
being written due to its cycles.

Changed the Tree.tsx to position the rootMember at the center of the
view, rather than at the top. This is because the rootMember may have
parents who need to be rendered above it. (I believe this is being
rendered above it, not sure for now)

Refactored the FamilyView to open NewMemberView when the button is
pressed, and close when the NewMemberView is open. Additionally, removed
the old joinFamily modal found here. Also removed the share family
button as this is no longer needed.

Removed an unused method in SocketContext now that the FamilyView has
been refactored